Simplify processStats() and processMems() by removing the error channels
The existence of /proc/pid/stat and /proc/pid/statm is checked only once, in GetProcessInfo()
As a result, these functions no longer generate errors.
This fixes a bug where these two functions would break when a process would die and the files would be unavailable.
In this case, it simply fails silently.